refactor: Engine 인터페이스 domain 단으로 이동
