+++ src/main/kotlin/domain/Engine.kt
@@ -1,4 +1,4 @@
-package util
+package domain
 
 fun interface Engine {
     fun getRandomNumber(): Int
